move to test_basic for client
